requirements.txt
aiogram==3.3.0
apscheduler==3.10.4
aiosqlite==0.19.0
python-dotenv==1.0.0
psutil==5.9.8 